+++ concordium-android-wallet-example/app/src/main/java/com/concordium/example/wallet/services/wallet_proxy/ProxyBackend.kt
@@ -1,0 +1,29 @@
+package com.concordium.example.wallet.services.wallet_proxy
+
+import com.concordium.sdk.responses.blocksummary.updates.queues.AnonymityRevokerInfo
+import com.concordium.sdk.responses.blocksummary.updates.queues.IdentityProviderInfo
+import retrofit2.Call
+import retrofit2.http.*
+import java.io.Serializable
+
+internal interface ProxyBackend {
+    @GET("v1/ip_info")
+    fun getIdentityProviderInfo(): Call<ArrayList<IdentityProvider>>
+}
+
+data class SubmissionData(
+    val submissionId: String
+)
+
+data class IdentityProvider(
+    val ipInfo: IdentityProviderInfo,
+    val arsInfos: Map<String, AnonymityRevokerInfo>,
+    val metadata: IdentityProviderMetaData
+) : Serializable
+
+data class IdentityProviderMetaData(
+    val icon: String,
+    val issuanceStart: String,
+    val support: String?,
+    val recoveryStart: String?
+) : Serializable