+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for android through an AAR artifact.
 - Added `ConcordiumHdWallet` class for deriving Concordium specific keys and randomness from a seed phrase.
 - Added `Identity` class with methods `createIdentityRequest` for creating identity requests and `createIdentityRecoveryRequest` for creating identity recovery requests.
+- Added `Credential` class exposing methods required to create, sign and sending a credential deployment transaction.
 
 ## 6.0.0
 - Added method `waitUntilFinalized` for waiting until a given transaction is finalized.